refactor(pack): refactor xdsl lan
