+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/modem/router/lan/index.js
@@ -1,0 +1,34 @@
+import angular from 'angular';
+import ngOvhTelecomUniverseComponents from '@ovh-ux/ng-ovh-telecom-universe-components';
+import '@uirouter/angularjs';
+import '@ovh-ux/ng-translate-async-loader';
+import 'angular-translate';
+import 'ovh-api-services';
+
+import controller from './lan.controller';
+import factory from './lan.factory';
+import template from './lan.html';
+
+const moduleName = 'XdslModemLan';
+
+angular
+  .module(moduleName, [
+    ngOvhTelecomUniverseComponents,
+    'ngTranslateAsyncLoader',
+    'pascalprecht.translate',
+    'ovh-api-services',
+    'ui.router',
+  ])
+  .controller('XdslModemLanCtrl', controller)
+  .factory('PackXdslModemLanObject', factory)
+  .run(/* @ngTranslationsInject:json ./translations */)
+  .run(
+    /* @ngInject */ ($templateCache) => {
+      $templateCache.put(
+        'app/telecom/pack/xdsl/modem/router/lan/pack-xdsl-modem-lan.html',
+        template,
+      );
+    },
+  );
+
+export default moduleName;
