+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/modem/router/lan/lan.html
